backend/requirements.txt
fastapi
uvicorn
requests
PyGithub
cachetools
